+++ src/lib/terrain/voxelmap/patch/patch-factory/merged/patch-factory-cpu.ts
@@ -1,6 +1,6 @@
-import * as THREE from '../../../../three-usage';
-import { type IVoxelMap, type VoxelsChunkSize } from '../../../terrain';
-import { VoxelsRenderableFactoryCpu } from '../../../voxelmap/voxelsRenderable/voxelsRenderableFactory/merged/cpu/voxels-renderable-factory-cpu';
+import * as THREE from '../../../../../three-usage';
+import { type IVoxelMap, type VoxelsChunkSize } from '../../../../terrain';
+import { VoxelsRenderableFactoryCpu } from '../../../voxelsRenderable/voxelsRenderableFactory/merged/cpu/voxels-renderable-factory-cpu';
 import { PatchFactoryBase, type GeometryAndMaterial } from '../patch-factory-base';
 
 class PatchFactoryCpu extends PatchFactoryBase {
